fix(imageviewer): explicit cast for indexOf narrowing conversion (clang-tidy)
Local clang-tidy (D:\LLVM, v22) over the M7-changed files surfaced one real
bugprone-narrowing-conversions finding: m_fileList.indexOf(path) returns
qsizetype assigned to int m_currentIndex. Cast explicitly (value is always a
small index). All other findings in changed files are style nits (pragma-once,
short names, magic-number 256) — advisory only, not gating.

Verified: build.ps1 Test -> 12/12 CTest green.

+++ src/imageviewer.cpp
@@ -76,7 +76,7 @@ void ImageViewer::setImage(const QString &path)
 
     const QFileInfo info(path);
     m_fileList = listImages(info.absolutePath());
-    m_currentIndex = m_fileList.indexOf(path);
+    m_currentIndex = static_cast<int>(m_fileList.indexOf(path));
 
     // Build the render pipeline state: tile grid over the full-res image, and
     // a Viewport fitted into the widget. The Widget no longer keeps its own
